Add 32bit support for driver (#99)
* initial changes to support building 32-bit

* adding scripts for building 32-bit and 64-bit versions of Driver on Windows

* updating tests where necessary to achieve full pass rate

* update workflow for building both 32-bit and 64-bit on Windows

* update workflow for building both 32-bit and 64-bit on Windows

* remove building installer from main build script

* 64-bit windows build script updates

* update workflow

* choose AWS library path based on bitness

* fixing warnings

* address code review comments

* updating README to reflect supported bitnesses

+++ libraries/rabbit/include/rabbit.hpp
@@ -340,7 +340,7 @@ namespace details {
 	
 	// Subtraction operator overload is required for std::distance
     template <typename OtherFunction, typename OtherIterator>
-    int64_t operator-(const transform_iterator<OtherFunction, OtherIterator>& other) const { return base() - other.base(); }
+    size_t operator-(const transform_iterator<OtherFunction, OtherIterator>& other) const { return base() - other.base(); }
 
     template <typename OtherFunction, typename OtherIterator>
     bool operator==(const transform_iterator<OtherFunction, OtherIterator>& other) const { return base() == other.base(); }
